Added system to verify invalid player assignments on remote clients when players join.
It is possible for a player to be assigned an object, but another client hasn't gotten that player's join message yet. This introduces a system to flag the invalid assignment and try assigning again when a new player joins. This is assuming that the player join message WILL happen after the OnDeserialization.
Verified Fix with CyanEmu by force setting the assignment array, having local player not be master, and adding remote players and seeing their pool objects auto assign on join.